Write attribute
