+++ src/app/clusters/basic-information/tests/BUILD.gn
@@ -20,13 +20,15 @@ import("${chip_root}/build/chip/chip_test_suite.gni")
 chip_test_suite("tests") {
   output_name = "libTestBasicInformationCluster"
 
-  test_sources = [ "TestBasicInformationCluster.cpp" ]
+  test_sources = [ "TestBasicInformationCluster.cpp",
+                   "TestBasicInformationReadWrite.cpp",  ]
 
   cflags = [ "-Wconversion" ]
 
   public_deps = [
     "${chip_root}/src/app/clusters/basic-information",
     "${chip_root}/src/app/clusters/testing",
+    "${chip_root}/src/app/server-cluster/testing",
     "${chip_root}/src/lib/core:string-builder-adapters",
     "${chip_root}/src/lib/support",
   ]
